Changed Console Location
After some testing and thought, I think it makes way more sense to make
the console full width in the main div on the bottom and just make the
height resizable. This will resolve some ugly issues when the window is
too small and will make the tool more usable at a smaller window width.