Firebase Integration
The share functionality should be working with Firebase as expected now.
I modified it to include the tab name and data. If there is a valid id
in the query string it will open a new tab and load the rule and data.
Since the query string is significantly reduced now, there is not much
need for the bitly integration, so it is removed for efficiency.